Parameter Make.Wasm_utils

type t

type tree

val initial_tree : version:Tezos_scoru_wasm.Wasm_pvm_state.version -> ?tree:tree -> ?ticks_per_snapshot:int64 -> ?max_reboots:Z.t -> ?from_binary:bool -> ?outbox_validity_period:int32 -> ?outbox_message_limit:Z.t -> string -> tree Lwt.t

val input_info : int32 -> Z.t -> Tezos_scoru_wasm.Wasm_pvm_state.input_info

val reveal_builtins : Tezos_scoru_wasm.Builtins.reveals

val eval_until_stuck : ?wasm_entrypoint:string -> ?reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> ?hooks:Tezos_scoru_wasm.Hooks.t -> ?write_debug:Tezos_scoru_wasm.Builtins.write_debug -> ?max_steps:int64 -> tree -> (Tezos_scoru_wasm.Wasm_pvm_errors.t * tree) Tezos_base.TzPervasives.tzresult Lwt.t

val eval_to_snapshot : ?wasm_entrypoint:string -> ?reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> ?hooks:Tezos_scoru_wasm.Hooks.t -> ?write_debug:Tezos_scoru_wasm.Builtins.write_debug -> ?max_steps:int64 -> tree -> tree Lwt.t

val eval_until_input_requested : ?wasm_entrypoint:string -> ?reveal_builtins:Tezos_scoru_wasm.Builtins.reveals option -> ?hooks:Tezos_scoru_wasm.Hooks.t -> ?write_debug:Tezos_scoru_wasm.Builtins.write_debug -> ?fast_exec:bool -> ?max_steps:int64 -> tree -> tree Lwt.t

val eval_until_input_or_reveal_requested : ?wasm_entrypoint:string -> ?hooks:Tezos_scoru_wasm.Hooks.t -> ?write_debug:Tezos_scoru_wasm.Builtins.write_debug -> ?fast_exec:bool -> ?max_steps:int64 -> tree -> tree Lwt.t

val set_sol_input : int32 -> tree -> tree Lwt.t

val set_protocol_migration_input : Tezos_scoru_wasm.Pvm_input_kind.protocol -> int32 -> tree -> tree Lwt.t

val set_info_per_level_input : ?migration_block:bool -> int32 -> tree -> tree Lwt.t

val set_eol_input : int32 -> Z.t -> tree -> tree Lwt.t

val set_raw_message : int32 -> Z.t -> string -> tree -> tree Lwt.t

val set_input_step : string -> int -> tree -> tree Lwt.t

val set_internal_message : int32 -> Z.t -> string -> tree -> tree Lwt.t

val set_inputs_step : ?migrate_to:Tezos_scoru_wasm.Pvm_input_kind.protocol -> (int32 -> Z.t -> 'a -> tree -> tree Lwt.t) -> 'a Tezos_base.TzPervasives.trace -> int32 -> tree -> tree Lwt.t

val set_full_input_step : ?migrate_to:Tezos_scoru_wasm.Pvm_input_kind.protocol -> string list -> int32 -> tree -> tree Lwt.t

val set_empty_inbox_step : ?migrate_to:Tezos_scoru_wasm.Pvm_input_kind.protocol -> int32 -> tree -> tree Lwt.t

val set_full_input_step_gen : ?migrate_to:Tezos_scoru_wasm.Pvm_input_kind.protocol -> (int32 -> Z.t -> 'a -> tree -> tree Lwt.t) -> 'a list -> int32 -> tree -> tree Lwt.t

val set_full_raw_input_step : ?migrate_to:Tezos_scoru_wasm.Pvm_input_kind.protocol -> string Tezos_base.TzPervasives.trace -> int32 -> tree -> tree Lwt.t

val eval_until_init : ?wasm_entrypoint:string -> tree -> tree Lwt.t

val eval_to_result : ?wasm_entrypoint:string -> ?write_debug:Tezos_scoru_wasm.Builtins.write_debug -> ?reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> ?hooks:Tezos_scoru_wasm.Hooks.t -> tree -> (tree * int64) Lwt.t

val pp_state : Stdlib.Format.formatter -> Tezos_scoru_wasm.Wasm_pvm_state.Internal_state.tick_state -> unit

val print_error_state : Tezos_scoru_wasm.Wasm_pvm_errors.t -> string

val check_error : ?expected_kind:Tezos_scoru_wasm_helpers_intf.Wasm_utils_intf.error_kind -> ?expected_reason:string -> Tezos_scoru_wasm.Wasm_pvm_errors.t -> bool

check_error kind reason error checks a Wasm PVM error error is of a given kind with a possible reason.

If kind is None, returns true.

If reason is None, it simply check the given kind, otherwise it actually check the reason in the error.

val is_stuck : ?step:Tezos_scoru_wasm_helpers_intf.Wasm_utils_intf.error_kind -> ?reason:string -> Tezos_scoru_wasm.Wasm_pvm_state.Internal_state.tick_state -> bool

val wrap_as_durable_storage : tree -> Tezos_webassembly_interpreter.Durable_storage.t Lwt.t

val has_stuck_flag : tree -> bool Lwt.t

val make_durable : (string * string) list -> Tezos_webassembly_interpreter.Durable_storage.t Lwt.t

val make_module_inst : version:Tezos_scoru_wasm.Wasm_pvm_state.version -> string list -> int32 -> Tezos_webassembly_interpreter.Instance.module_inst Tezos_webassembly_interpreter.Instance.ModuleMap.t * Tezos_webassembly_interpreter.Instance.module_key * Tezos_webassembly_interpreter.Host_funcs.registry

val retrieve_memory : Tezos_webassembly_interpreter.Instance.module_inst Tezos_webassembly_interpreter.Instance.ModuleMap.t -> Tezos_webassembly_interpreter.Partial_memory.memory Lwt.t

module Ctx : Tezos_tree_encoding.Encodings_util.S with type Tree.tree = tree with type t = t

module Tree_encoding_runner : Tezos_tree_encoding.Runner.S with type tree = tree

module Wasm : Tezos_scoru_wasm.Wasm_pvm_sig.S with type tree = tree

module Wasm_fast : Tezos_scoru_wasm.Wasm_pvm_sig.S with type tree = tree