Module Wasm_utils.Tree_encoding_runner

type tree = tree

val encode : 'a Tezos_tree_encoding.t -> 'a -> tree -> tree Lwt.t

encode enc x tree encodes a value x using the encoder enc into the provided tree.

val decode : 'a Tezos_tree_encoding.t -> tree -> 'a Lwt.t

decode enc x tree decodes a value using the encoder enc from the provided tree.