Module Octez_smart_rollup_wasm_debugger_lib.Timing_buffer

type t

Timing buffer which records elapsed time associated with each line written to it

val create : unit -> t

Create a new timing buffer.

val add_message : t -> Stdlib.out_channel -> string -> unit

Add a message. If there is something to output it will be written to the provided channel.

val flush_buffer : t -> Stdlib.out_channel -> unit

Flush the buffer contents to the given channel.