Module Octez_smart_rollup_wasm_debugger_lib.Wasm_debugger

val read_kernel_from_file : Lwt_io.file_name -> (string * bool, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

read_kernel_from_file path returns the contents stored in path, along with a boolean hinting if the contents is expected to be binary (true) or text (false).

val parse_binary_module : string -> string -> Tezos_webassembly_interpreter.Ast.module_ Lwt.t

val typecheck_module : Tezos_webassembly_interpreter.Ast.module_ -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val import_pvm_host_functions : version:Tezos_scoru_wasm.Wasm_pvm_state.version -> unit -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val link : Tezos_webassembly_interpreter.Ast.module_ -> Tezos_webassembly_interpreter.Instance.extern list Tezos_base.TzPervasives.tzresult Lwt.t

val check_kernel : binary:bool -> name:string -> Tezos_scoru_wasm.Wasm_pvm_state.version -> string -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

module Make (Wasm : Tezos_scoru_wasm_helpers_intf.Wasm_utils_intf.S) : sig ... end